Add func alert(_: UIAlertController?) -> Alert
